Refactor obj/bundle generation utils
This commit refactors data object and data bundle generation
utilities used in the compliance testing suite to a new testing
superclass, :class:`ga4gh.dos.test.DataObjectServiceTest`. This
helps reduce code repetition between the server tests and the
compliance tests, and also lays the foundation for allowing DOS
implementations to also test themselves with the server tests.

This commit also implements some typing improvements in the
data object and data bundle generation methods.
